Change index.pickle to contain 50 descriptors
I realized that if the user draws too small a line, it won't require as
my fourier transforms. Basically, I think pixelsDrawn = fourier
descriptors I can use because FAISS requires it to be the same size.
I'll also try to blur the image before putting it into Canny because
these edges are almost too defined. I want loose, broken up edges so
that there are a lot of edges to match the user's drawing with.